+++ CHANGELOG.md
@@ -46,6 +46,10 @@ For each minor and major version, there is a corresponding [milestone on GitHub]
 
 ## v0.3.0 (2025-??-??)
 
+### New features
+
+- The `$name`s of kinds are configurable now.
+
 ### Fixed bugs
 
 - The logic to ensure that types are not created multiple times needs to check that the kind of the types is the same. Otherwise a collision of duplicated identifiers of types needs to be reported.